PowerProfileService: clean up UI dependencies
Services are low-level classes that don't need to know about anything
UI-related, such as translations, icons, or colors. At the same time,
the string<->enum conversion is internal to the service and not
something the UI should care about.

Ref: #375
